Create version_check_based_on_csv_lists.py
This script can extract the individual IP addresses/domains from a csv table and pass them to `scan-citrix-netscaler-version.py`. The results are compared directly to determine whether the systems are vulnerable to current CVEs or not. All results are supplemented with the other available information and written to a new CSV. 

I didn't want to mention the original script, as this is only an optional solution for certain cases.